Move pages to src/pages
