--- pages/profiles/[geoIdOrCountrySlug]/[analysisSlug]/index.js
@@ -1,3 +1,0 @@
-import Analysis from '../../../../src/pages/Analysis';
-
-export default Analysis;
