--- pages/terms.js
@@ -1,3 +1,0 @@
-import Legal from '../src/pages/Legal';
-
-export default Legal;
