+++ src/pages/_app.js
@@ -1,13 +1,15 @@
 import 'cross-fetch/polyfill';
 
 import React from 'react';
+
 import App from 'next/app';
+import ApolloClient from 'apollo-boost';
+import { ApolloProvider } from '@apollo/react-hooks';
+
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { ThemeProvider } from '@material-ui/core/styles';
 
-import ApolloClient from 'apollo-boost';
-import { ApolloProvider } from '@apollo/react-hooks';
-import Theme from '../src/Theme';
+import theme from '../theme';
 
 const client = new ApolloClient({
   uri: 'https://graphql.takwimu.africa/graphql'
@@ -27,7 +29,7 @@ export default class MyApp extends App {
 
     return (
       <ApolloProvider client={client}>
-        <ThemeProvider theme={Theme}>
+        <ThemeProvider theme={theme}>
           <CssBaseline />
           <Component {...pageProps} />
         </ThemeProvider>
