+++ src/pages/methodology.js
