+++ src/pages/privacy.js
@@ -1,0 +1,3 @@
+import Legal from './legal';
+
+export default Legal;
