+++ src/pages/profiles/[geoIdOrCountrySlug]/[analysisSlug]/index.js
@@ -1,0 +1,3 @@
+import Analysis from '../../../../components/AnalysisPage';
+
+export default Analysis;
